Demos: create exclusive file and interface for demos
